add processes for tau_hwp
